+++ apiphany-httpclient5/src/test/java/org/apiphany/http/ApacheHC5ClientsTest.java
@@ -3,6 +3,9 @@ package org.apiphany.http;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 
+import java.io.IOException;
+
+import org.apache.hc.client5.http.CircularRedirectException;
 import org.apache.hc.client5.http.config.RequestConfig;
 import org.apiphany.client.ClientProperties;
 import org.junit.jupiter.api.Test;
@@ -31,5 +34,27 @@ class ApacheHC5ClientsTest {
 		RequestConfig requestConfig = ApacheHC5Clients.createRequestConfig(properties);
 
 		assertThat(requestConfig.isRedirectsEnabled(), equalTo(false));
+	}
+
+	@Test
+	void shouldIdentifyCircularRedirectException() throws CircularRedirectException {
+		CircularRedirectException exception = new CircularRedirectException("circular redirect");
+
+		assertThat(ApacheHC5Clients.isCircularRedirectException(exception), equalTo(true));
+	}
+
+	@Test
+	void shouldIdentifyCircularRedirectExceptionWrappedInCauseChain() {
+		CircularRedirectException circular = new CircularRedirectException("circular redirect");
+		IOException wrapped = new IOException("circular redirect", circular);
+
+		assertThat(ApacheHC5Clients.isCircularRedirectException(wrapped), equalTo(true));
+	}
+
+	@Test
+	void shouldNotIdentifyNonCircularRedirectException() {
+		IOException exception = new IOException("some error");
+
+		assertThat(ApacheHC5Clients.isCircularRedirectException(exception), equalTo(false));
 	}
 }